Deleted "also", meant to do that moments ago.
Thing changed
         Explanation
--------------------------------------------------------------------
The default is base = 16, which means…
	Comma is grammatically necessary.

To make the interpreter display bitvectors in decimal (base 10), type :set base=10. To change back to hexadecimal, type :set base=16.
	Commas helpful for parsing verb (vs. noun) meaning of “type”.

For example, the number 10
	Typo, needs lowercase t.

In the last section that you'll see with :b are Cryptol's symbols.
	Added “that” for easier parsing on first read (clarifying “you’ll see...” modifies “section” rather than “In the last section modifying “see”).

The interpreter supports TAB completion; that is, pressing TAB will display all available symbols.
	Semicolon to indicate introduction of second clause, rather than appositive.  Remove “of” for grammar.

Though, in the same way, now the value of it has become [1, 2, 3].
	“Also” doesn’t fit, since there are not two values bound at once.  Perhaps “accordingly” or “in the same way” is meant.

and program understanding; however, it can cause
	Semicolon to separate independent clauses.

Here we see that this module provides one type synonym and three symbols.  (If your result does not match, try reloading the module.)
	The previous exercise may have changed x and y so they don’t appear in the output list shown.

Changed to "Namespace elements of a module name directly correlate with directory structure; e.g., this module labs::Interpreter::Interpreter maps directly to the file at labs/Interpreter/Interpreter.md.  Thus in order for the module to be found (essentially, finding the base directory for that path), either"
	This is an attempt to add one logic step between the description of path and file names associated with module names, on the one hand, and the “thus” clause that follows, which lists two options to make sure the module is found, on the other hand.  As it stood, I think I would not readily see that the fact “module name=path and file name” (essentially) implies that to find the module, the interpreter needs to be opened in a particular directory, or the CRYPTOLPATH environment variable must contain the base directory of the module.  For those of us who don’t automatically think, “As usual, pathnames referenced in code usually will cause the software to search for the beginning of that path in the current directory or in a base directory listed in an environment variable,” I think this phrase may be helpful.

Changed to "When you type `:e`, Cryptol will open the current file or module in a text editor, and will then reload the file into the interpreter when the editor is closed."
	Clarifying the two uses of the word “load” by replacing one.