Basics-pattern-matching-suggested-clarification
Basics.md, suggestions for pattern matching examples.
Working from master branch.  8 Sep 2020 6:36 p.m.

The pattern-matching explanation is not as transparent as it could be.  I think the person who put it together was trying to make the names meaningful by calling `(a, b)` by the identifier `ab` and giving it the value `(0xa, 0xb)`, so that the assignment `let (a, b) = ab` would give `a` the value `0xa` and b the value `0xb`.

I, however, had to stare at the variable assignments and explicitly think through the distinctions, “Here `a` and `b` are part of a single identifier; they are not being multiplied; here they are hex digits; and here they are separate identifiers.”  Notice that this struggle is really a side issue to the question of what pattern matching does.

I think it would be better, then, to start off with a completely straightforward example, using literal values that don’t look like variables, and not reusing characters in a way that might introduce confusion (confusion unrelated to the concept being taught, that it).  Hence the suggestion and the extra verbiage I offer here.