Basic--clear up type/value language, mark quote.
Basics.md, suggestions for explanation and examples about type variable and value variable passing.
Working from master branch.  9 Sep 2020 2:21 p.m.

Submitted for your consideration:

First, about strong typing:  I believe the lab quotes a Wikipedia page, so this should be designated as a quotation (“generally refers...programming errors”), maybe as a block quotation.

Second, I am looking at the section of Basics.md beginning with the sentence, “Now, because there are two classes of variables, there are distinct ways of passing type variables and value variables to a function.”

The issue is this:  There is some ambiguity about the terminology used regarding passing a parameter.  There are phrases like “passing a variable” and “passing a value,” or passing a value “as a value argument” (and presumably passing a value of a type argument).  All or nearly all of these phrases show up in the Programming Cryptol manual, beginning of section 1.20 Type vs. value variables, pp. 23 bottom--24 top.

I would like this clarified.  Hence the changes I have suggested.  I am taking the position that “passing a variable” means the same thing as “passing the value of that variable,” or else means the same thing but at a higher, more abstract level.  So that is how I explain it in my suggested changes, and if anyone would like to tell me differently, I will welcome that.