Basics--note to recover overwritten code snippets
Basics.md, just telling students what to do if they save over a code snippet and want to recover it.

Working from master branch.  9 Sep 2020 3:20 p.m.

The suggested change is self-explanatory, just a note to students that many would find obvious (so perhaps unnecessary)--but not me when I took the class, for some reason; I always copied the snippet in place so I'd have a backup.  Also, for students unfamiliar with GitHub, it may serve as an encouragement to browse there a little.